requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Chess library for position evaluation
python-chess>=1.999

# API requests for Lichess
requests>=2.31.0

# Data processing
python-dateutil>=2.8.2
pyyaml>=6.0

# Optional: for advanced ML models
# xgboost>=2.0.0
# lightgbm>=4.0.0

# Optional: for Lichess API (recommended for fetching real games)
# berserk>=0.11.0  # Official Lichess API client - uncomment to use

# Development dependencies
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Visualization for experiments
matplotlib>=3.7.0

